adding tags an ability to target live and staging.
* deployments ( teardown ) target's should only ever target staging.
Teardown should never set is_promote_lb_tg let the default handle it.

* replacement for move dns pointer jenkins job should promote current
staging to live tg and updates is live tag.  Also demotes live to
staging and updates tag.  Simply set the is_promote_lb_tg to true.

* needed to create the tag:Stack on one of the lbs as it was missing.
Maybe tag:subDomain is a better tag name for the lb.  Needs to be
handled by the alb module that controls the state of the lb.