[ALS-5082] Persist database: Flyway (#126)
* [ALS-5082] Remove unused code

This code is tracked by version control and I would prefer not to have dead code within the codebase.

* [ALS-5082] Update MySQL connector version in Wildfly module config

The MySQL connector version in the Wildfly configuration file has been updated from 5.1.38 to 8.0.33. This is necessary for keeping with the latest stable version for improved performance and security.

* [ALS-5082] Add initial roles, privileges, resources, and create initial tables for database

The commit includes SQL scripts to insert initial roles and privileges in the 'auth' database and to create the initial tables required for 'picsure' and 'auth' databases. Additionally, the commit inserts resources into the 'picsure' database. These updates initialize the foundational data structures needed for the application's function.

* Refactor DNS records management in terraform

Removed deprecated comments, unused code and the reliance on Jenkins and AWS CLI for managing route53 DNS records in the route53-template.tf file. The refactoring now utilizes Terraform natively, directly handling AWS Route53 changes and improving code clarity.

* [ALS-5082] Update Terraform configuration and application to use persisted database, remove unnecessary vars, use DB username/host/password vars, add resource IDs and optimize AWS instance size and volume. This change allows us to have a persistent database for our application, which retains data across sessions and improves the overall performance and cost-effectiveness of the AWS instance. Various unnecessary variables have been obliterated, further streamlining the configuration process.

* [ALS-5082] Remove resource-registration.sql