[ALS-6198] Convert picsure resource id to a variable (#134)
* [ALS-6198] Convert picsure resource id to a variable

Some request are being sent to the API with the auth-hpds UUID. This is likely a result of this UUID being used in some request. We need to ensure open or auth queries the appropriate HPDS type.

* [ALS-6198] Update AWS instance type in wildfly-instance Terraform file

Changed the AWS instance type for wildfly-ec2 resource in Terraform configuration file. It's been updated from "m5.large" to "m5.2xlarge" to enhance the server capacity and handle more traffic efficiently.