Update S3 path locations in scripts and configs
Removed the dependency on specific git hash from S3 paths in various Terraform scripts and config files, and set them to point directly to files without specifying a git hash. It allows for consistent file paths that don't depend on a specific git commit, simplifying the maintenance process and making the scripts more adaptable to changes in the repository. Also added a new 'open-hpds-docker.sh' script in the git to perform certain initialization tasks.