Add remote debug capability to PSAMA script
The script `psama-docker.sh` was modified to include an additional parameter that determines whether remote debugging is available or not. If debugging is enabled, corresponding JAVA_OPTS are added. Also, adjustments in the `wildfly-docker.sh` and `wildfly-user_data.sh` scripts were done in order to maintain consistent argument numbering across all scripts.