[ALS-7051] Deploy Data Dictionary API (#152)
* Update HTTPD ec2 user script for new UI

* Reroute to localhost

* Update httpd-vhosts.conf

* deprecated dictionary resource
* dictionary will be run on the wildfly instance

* user-script and container run script

* iam updates

* removing dictionary resources from outputs and route 53

* Update wildfly-iam.tf

* add containers/application condition

* fixing up container run

# Both containers should start up now
* api looks to be running and staying up
* the weights container is in a restart loop due to not being able to connect to the db which is expected

* New UI - updates (#150)

* Update zip name

* Update tar name

* Fix routing

---------

Co-authored-by: James <Jamestp19@gmail.com>

* Remove dictionary weights image and container

Eliminated the loading and running of the dictionary weights Docker image and container alongside the related S3 copy command. This simplifies the deployment script by focusing solely on the dictionary API. We will deploy the weights service on Jenkins as it is a temporary container.

* Create dictionary resource and authorization rules

Added SQL scripts to create a dictionary resource and an access rule for it. The dictionary resource is defined with specific endpoint details, while the corresponding access rule allows routing requests to this resource. This setup is essential for enabling API interactions and ensuring proper access control.

* Remove unnecessary .env file and comments from script

Removed the creation of a blank .env file and redundant comments in the `dictionary-docker.sh` script. Removed an unnecessary S3 copy command for weights.csv and ensured the Docker container is properly restarted by stopping and removing the existing one before running a new instance.

* Add IAM permissions for new PostgreSQL secret

Extended IAM policy to include permissions for accessing the PostgreSQL user secret. Introduced a new variable for the PostgreSQL secret name in the variables file.

* Fix incorrect parameter in dictionary resource SQL script

Changed 'dictResourceUUID' to '@dictResourceUUID' in the SQL insert statement to properly reference the variable. This ensures the dictionary resource is created with the correct UUID parameter.

---------

Co-authored-by: James <Jamestp19@gmail.com>
Co-authored-by: TDeSain <thomas.n.desain@gmail.com>
Co-authored-by: TDeSain <thomas_desain@hms.harvard.edu>