Deploy new UI in place (#161)
* Add script to run HTTPD Docker container
Created a new script `httpd-docker.sh` to handle running the HTTPD Docker container and updated `httpd-user_data.sh` to call this script. This change aims to modularize the code and improve maintainability by separating the concerns of the HTTPD Docker setup.

* Stop and remove existing 'httpd' container before run
Added commands to stop and remove the existing 'httpd' container in the httpd-docker.sh script. This ensures that the docker run command starts a fresh instance without conflicts from previous containers.