ALS-7359 (#162)
* User-script updates
# Per request
* moving yum updates to end of user-scripts
* stopping splunk forwarder before enable boot-start

* Add support for picsure-dictionary environment configuration

This commit includes updates to allow fetching the picsure-dictionary.env file from S3 and using it in the Docker container. The IAM policy and the dictionary-docker.sh script have been modified accordingly.

* Add support for picsure-dictionary environment configuration

This commit includes updates to allow fetching the picsure-dictionary.env file from S3 and using it in the Docker container. The IAM policy and the dictionary-docker.sh script have been modified accordingly.